create group prefix values and rename ACLUtil test method
